% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlieSkillsEngland-package.R
\docType{package}
\name{charlieSkillsEngland-package}
\alias{charlieSkillsEngland}
\alias{charlieSkillsEngland-package}
\title{charlieSkillsEngland: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
