% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-functions.R
\name{complex_function}
\alias{complex_function}
\title{complex_function}
\usage{
complex_function(number, exponent)
}
\arguments{
\item{number}{A Number}

\item{exponent}{The power the number is put to the power of}
}
\description{
complex_function
}
\examples{
complex_function(10, 3)
}
