% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-functions.R
\name{data_function}
\alias{data_function}
\title{data_function}
\usage{
data_function(data)
}
\arguments{
\item{number}{A Number}

\item{exponent}{The power the number is put to the power of}
}
\description{
data_function
}
\examples{
data_function(10, 3)
}
