% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{devolved_skills_departments}
\alias{devolved_skills_departments}
\title{Devolved Nations Skills Departments}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
devolved_skills_departments
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Devolved Nations Skills Departments
}
\keyword{datasets}
