% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-functions.R
\name{to_the_power}
\alias{to_the_power}
\title{to_the_power is a silly function}
\usage{
to_the_power(number, exponent)
}
\arguments{
\item{number}{A Number}

\item{exponent}{The power the number is put to the power of}
}
\description{
to_the_power is a silly function
}
\examples{
to_the_power(10, 3)
}
